Refactor containermetrics reporting and envoy ceration for better separation of duties
