+++ depot/containerstore/containerstorefakes/fake_containerstore.go
@@ -1119,36 +1119,6 @@ func (fake *FakeContainerStore) UpdateReturnsOnCall(i int, result1 error) {
 func (fake *FakeContainerStore) Invocations() map[string][][]interface{} {
 	fake.invocationsMutex.RLock()
 	defer fake.invocationsMutex.RUnlock()
-	fake.cleanupMutex.RLock()
-	defer fake.cleanupMutex.RUnlock()
-	fake.createMutex.RLock()
-	defer fake.createMutex.RUnlock()
-	fake.destroyMutex.RLock()
-	defer fake.destroyMutex.RUnlock()
-	fake.getMutex.RLock()
-	defer fake.getMutex.RUnlock()
-	fake.getFilesMutex.RLock()
-	defer fake.getFilesMutex.RUnlock()
-	fake.initializeMutex.RLock()
-	defer fake.initializeMutex.RUnlock()
-	fake.listMutex.RLock()
-	defer fake.listMutex.RUnlock()
-	fake.metricsMutex.RLock()
-	defer fake.metricsMutex.RUnlock()
-	fake.newContainerReaperMutex.RLock()
-	defer fake.newContainerReaperMutex.RUnlock()
-	fake.newRegistryPrunerMutex.RLock()
-	defer fake.newRegistryPrunerMutex.RUnlock()
-	fake.remainingResourcesMutex.RLock()
-	defer fake.remainingResourcesMutex.RUnlock()
-	fake.reserveMutex.RLock()
-	defer fake.reserveMutex.RUnlock()
-	fake.runMutex.RLock()
-	defer fake.runMutex.RUnlock()
-	fake.stopMutex.RLock()
-	defer fake.stopMutex.RUnlock()
-	fake.updateMutex.RLock()
-	defer fake.updateMutex.RUnlock()
 	copiedInvocations := map[string][][]interface{}{}
 	for key, value := range fake.invocations {
 		copiedInvocations[key] = value
