+++ depot/containerstore/containerstorefakes/fake_cred_handler.go
@@ -324,14 +324,6 @@ func (fake *FakeCredentialHandler) UpdateReturnsOnCall(i int, result1 error) {
 func (fake *FakeCredentialHandler) Invocations() map[string][][]interface{} {
 	fake.invocationsMutex.RLock()
 	defer fake.invocationsMutex.RUnlock()
-	fake.closeMutex.RLock()
-	defer fake.closeMutex.RUnlock()
-	fake.createDirMutex.RLock()
-	defer fake.createDirMutex.RUnlock()
-	fake.removeDirMutex.RLock()
-	defer fake.removeDirMutex.RUnlock()
-	fake.updateMutex.RLock()
-	defer fake.updateMutex.RUnlock()
 	copiedInvocations := map[string][][]interface{}{}
 	for key, value := range fake.invocations {
 		copiedInvocations[key] = value
