+++ depot/transformer/faketransformer/fake_transformer.go
@@ -111,8 +111,6 @@ func (fake *FakeTransformer) StepsRunnerReturnsOnCall(i int, result1 ifrit.Runne
 func (fake *FakeTransformer) Invocations() map[string][][]interface{} {
 	fake.invocationsMutex.RLock()
 	defer fake.invocationsMutex.RUnlock()
-	fake.stepsRunnerMutex.RLock()
-	defer fake.stepsRunnerMutex.RUnlock()
 	copiedInvocations := map[string][][]interface{}{}
 	for key, value := range fake.invocations {
 		copiedInvocations[key] = value
