+++ fakes/fake_client.go
@@ -1242,40 +1242,6 @@ func (fake *FakeClient) VolumeDriversReturnsOnCall(i int, result1 []string, resu
 func (fake *FakeClient) Invocations() map[string][][]interface{} {
 	fake.invocationsMutex.RLock()
 	defer fake.invocationsMutex.RUnlock()
-	fake.allocateContainersMutex.RLock()
-	defer fake.allocateContainersMutex.RUnlock()
-	fake.cleanupMutex.RLock()
-	defer fake.cleanupMutex.RUnlock()
-	fake.deleteContainerMutex.RLock()
-	defer fake.deleteContainerMutex.RUnlock()
-	fake.getBulkMetricsMutex.RLock()
-	defer fake.getBulkMetricsMutex.RUnlock()
-	fake.getContainerMutex.RLock()
-	defer fake.getContainerMutex.RUnlock()
-	fake.getFilesMutex.RLock()
-	defer fake.getFilesMutex.RUnlock()
-	fake.healthyMutex.RLock()
-	defer fake.healthyMutex.RUnlock()
-	fake.listContainersMutex.RLock()
-	defer fake.listContainersMutex.RUnlock()
-	fake.pingMutex.RLock()
-	defer fake.pingMutex.RUnlock()
-	fake.remainingResourcesMutex.RLock()
-	defer fake.remainingResourcesMutex.RUnlock()
-	fake.runContainerMutex.RLock()
-	defer fake.runContainerMutex.RUnlock()
-	fake.setHealthyMutex.RLock()
-	defer fake.setHealthyMutex.RUnlock()
-	fake.stopContainerMutex.RLock()
-	defer fake.stopContainerMutex.RUnlock()
-	fake.subscribeToEventsMutex.RLock()
-	defer fake.subscribeToEventsMutex.RUnlock()
-	fake.totalResourcesMutex.RLock()
-	defer fake.totalResourcesMutex.RUnlock()
-	fake.updateContainerMutex.RLock()
-	defer fake.updateContainerMutex.RUnlock()
-	fake.volumeDriversMutex.RLock()
-	defer fake.volumeDriversMutex.RUnlock()
 	copiedInvocations := map[string][][]interface{}{}
 	for key, value := range fake.invocations {
 		copiedInvocations[key] = value
