+++ initializer/fakes/fake_cert_pool_retriever.go
@@ -84,8 +84,6 @@ func (fake *FakeCertPoolRetriever) SystemCertsReturnsOnCall(i int, result1 *x509
 func (fake *FakeCertPoolRetriever) Invocations() map[string][][]interface{} {
 	fake.invocationsMutex.RLock()
 	defer fake.invocationsMutex.RUnlock()
-	fake.systemCertsMutex.RLock()
-	defer fake.systemCertsMutex.RUnlock()
 	copiedInvocations := map[string][][]interface{}{}
 	for key, value := range fake.invocations {
 		copiedInvocations[key] = value
